ALMOOSA-848 critical-care ui design

+++ src/components/header.vue
@@ -242,7 +242,7 @@ export default {
       userService.setSelectedLayout(this.layoutType);
     },
     logout() {
-      this.$root.$refs.appointmentModule.destroyObjects();
+      this.$root.$refs.appointmentModule && this.$root.$refs.appointmentModule.destroyObjects();
       this.removeUserInfo();
       this.$messaging.deleteToken();
       this.navigateTo({ name: "Login Dashboard" });
